Fix access to user selected files
`ChordProMac` is _sandboxed_ and that will restrict access to files outside of the application. It a good thing but makes it much harder to develop a macOS application. I can just 'un-sandbox' the application and call it the day but that will make the wrapper not future-proof and is lazy coding :-)