macOS Classic dndhandler in Swift
An experimental Swift alternative for the current dndhandler that is written in C++.

It is completely optional; by default, `dndhandler.cpp` will be used.

To compile with the experimental Swift version:

	make DNDHANDLER=swift

Minimum target is macOS Catalina; however, it is only tested on Monterey, Ventura and Sonoma on Apple Silicon.

I've changed the identifier of the SwiftUI wrapper to `org.chordpro.chordpromac` because it must be unique and `org.chordpro.chordpro` is already for the Classic version. Cleaned-up Info.plist for both versions so you can see in the info if it is Classic or SwiftUI.