(macOS) Remove SwiftUI
I’ve decided to drop the development on the Mac SwiftUI GUI for ChordPro. While looking great; it took a lot of effort to support the Mac with my coding while the other platforms use wxWidgets for its GUI and its code is fully Open Source and native to the ChordPro core.

Over the last few months, the multi-platform wxChordPro GUI became so much better; also for the Classic Mac.

Me, as a Mac snob, I actually like it.

Bye bye SwiftUI!

The ChordPro journey goes on!

Nick