(Wx) New dndhandler for macOS
Run `dndhandler` as background process and launch a separate `wxchordpro` process for each file that is dragged/dropped or opened by double-click in the Finder.

This will allow multiple *ChordPro* windows to be open at the same time and fixes the issue that the `dndhandler` only worked once at startup.

When drag/drop or clicking a song that is already open, that window will be brought in front instead of opening a new window; standard macOS behaviour.

When the last window is closed, the background process will be terminated; so no hidden processes when no window is open.

Known issues:

- Each window will have its own icon in the Dock
- When starting *ChordPro* or opening a new document in its own window, the dock icon does some strange bouncing (once)

Tested on macOS Ventura and Sequoia