diagrams:draw dots and fingers always
--some pdf readers under linux like okular,
when no color has been set for dots in diagrams
they probably paint them transparent and thus are
not shown at all. That was occuring when a cho
file has set to draw its fingers "below" in the
chord diagrams. The previous implementation did
not set a default color to fall back for dots.
By setting a default color for dots to fallback
everything is shown correctly under
okular even the fingers below the diagram.