Update macOS application icon
On macOS Tahoe (26), the application icon is not allowed to go out of its squarish boundary anymore; like the current ChordPro icon does. If it does, it will get an ugly grey frame around it.

So I updated the icon. Not perfect because the squarish shape is actually changed a bit as well... Nothing I can do about that and at least it looks better now on Tahoe and still good on older versions.