Update config.schema
Removed `additionalProperties` in places that don't work correctly. Due to segmenting some schemas to reuse shared subschema structure, this causes `additionalProperties` to not function in the intended way. In place of `additionalProperties`, `unevaluatedProperties` should be used, but it's not supported in `Draft 07` of JSON Schema. Instead, `$comment`s were left as a note for the future.